Fix Delta 2 message parsing by introducing partial decoding (#118)
It seems that some Delta 2 devices do not decode the whole message byte stream for some messages as reported in #116. Parsing code suggests that the messages are composed of base message and multiple extensions depending on the length of the message.

This PR adds partial decoding based on the length of the incoming message. This implementation is not super efficient as it goes down from the full size and removes one struct format string each iteration while calling struct.calcsize each time, but it is then memorized for that length so it should be fine.