Add region to config form (#120)
We are using api.ecoflow.com to get the user id, but it seems that this redirects to the region the user makes request from which should work for most of the users, however in #113 it was reported that the account was created for a different region so this PR adds a selector for region to login to.

It seems that the user id is also not the same depending on which region you choose - I do have accounts for both US and EU regions and it is actually different.